requirements.txt
arch==8.0.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
statsmodels==0.14.6
threadpoolctl==3.6.0
tqdm==4.67.3
tzdata==2025.3
xgboost==3.2.0
